move nms functions to their own files
