+++ tfjs-core/src/ops/image_ops.ts
@@ -15,150 +15,13 @@
  * =============================================================================
  */
 
-import {nonMaxSuppressionV3, nonMaxSuppressionV5} from '../backends/non_max_suppression_impl';
 import {ENGINE, ForwardFunc} from '../engine';
-import {Tensor, Tensor1D, Tensor2D, Tensor4D} from '../tensor';
-import {NamedTensorMap} from '../tensor_types';
+import {Tensor1D, Tensor2D, Tensor4D} from '../tensor';
 import {convertToTensor} from '../tensor_util_env';
 import {TensorLike} from '../types';
 import * as util from '../util';
-import {nonMaxSuppSanityCheck} from './nonmax_util';
 
 import {op} from './operation';
-export {nonMaxSuppression} from './non_max_suppression';
-
-/**
- * Performs non maximum suppression of bounding boxes based on
- * iou (intersection over union).
- *
- * @param boxes a 2d tensor of shape `[numBoxes, 4]`. Each entry is
- *     `[y1, x1, y2, x2]`, where `(y1, x1)` and `(y2, x2)` are the corners of
- *     the bounding box.
- * @param scores a 1d tensor providing the box scores of shape `[numBoxes]`.
- * @param maxOutputSize The maximum number of boxes to be selected.
- * @param iouThreshold A float representing the threshold for deciding whether
- *     boxes overlap too much with respect to IOU. Must be between [0, 1].
- *     Defaults to 0.5 (50% box overlap).
- * @param scoreThreshold A threshold for deciding when to remove boxes based
- *     on score. Defaults to -inf, which means any score is accepted.
- * @return A 1D tensor with the selected box indices.
- */
-/** @doc {heading: 'Operations', subheading: 'Images', namespace: 'image'} */
-/** This is the async version of `nonMaxSuppression` */
-async function nonMaxSuppressionAsync_(
-    boxes: Tensor2D|TensorLike, scores: Tensor1D|TensorLike,
-    maxOutputSize: number, iouThreshold = 0.5,
-    scoreThreshold = Number.NEGATIVE_INFINITY): Promise<Tensor1D> {
-  const $boxes = convertToTensor(boxes, 'boxes', 'nonMaxSuppressionAsync');
-  const $scores = convertToTensor(scores, 'scores', 'nonMaxSuppressionAsync');
-
-  const inputs = nonMaxSuppSanityCheck(
-      $boxes, $scores, maxOutputSize, iouThreshold, scoreThreshold);
-  maxOutputSize = inputs.maxOutputSize;
-  iouThreshold = inputs.iouThreshold;
-  scoreThreshold = inputs.scoreThreshold;
-
-  const boxesAndScores = await Promise.all([$boxes.data(), $scores.data()]);
-  const boxesVals = boxesAndScores[0];
-  const scoresVals = boxesAndScores[1];
-
-  const res = nonMaxSuppressionV3(
-      boxesVals, scoresVals, maxOutputSize, iouThreshold, scoreThreshold);
-  if ($boxes !== boxes) {
-    $boxes.dispose();
-  }
-  if ($scores !== scores) {
-    $scores.dispose();
-  }
-  return res;
-}
-
-/**
- * Performs non maximum suppression of bounding boxes based on
- * iou (intersection over union).
- *
- * This op also supports a Soft-NMS mode (c.f.
- * Bodla et al, https://arxiv.org/abs/1704.04503) where boxes reduce the score
- * of other overlapping boxes, therefore favoring different regions of the image
- * with high scores. To enable this Soft-NMS mode, set the `softNmsSigma`
- * parameter to be larger than 0.
- *
- * @param boxes a 2d tensor of shape `[numBoxes, 4]`. Each entry is
- *     `[y1, x1, y2, x2]`, where `(y1, x1)` and `(y2, x2)` are the corners of
- *     the bounding box.
- * @param scores a 1d tensor providing the box scores of shape `[numBoxes]`.
- * @param maxOutputSize The maximum number of boxes to be selected.
- * @param iouThreshold A float representing the threshold for deciding whether
- *     boxes overlap too much with respect to IOU. Must be between [0, 1].
- *     Defaults to 0.5 (50% box overlap).
- * @param scoreThreshold A threshold for deciding when to remove boxes based
- *     on score. Defaults to -inf, which means any score is accepted.
- * @param softNmsSigma A float representing the sigma parameter for Soft NMS.
- *     When sigma is 0, it falls back to nonMaxSuppression.
- * @return A map with the following properties:
- *     - selectedIndices: A 1D tensor with the selected box indices.
- *     - selectedScores: A 1D tensor with the corresponding scores for each
- *       selected box.
- */
-/** @doc {heading: 'Operations', subheading: 'Images', namespace: 'image'} */
-function nonMaxSuppressionWithScore_(
-    boxes: Tensor2D|TensorLike, scores: Tensor1D|TensorLike,
-    maxOutputSize: number, iouThreshold = 0.5,
-    scoreThreshold = Number.NEGATIVE_INFINITY,
-    softNmsSigma = 0.0): NamedTensorMap {
-  const $boxes = convertToTensor(boxes, 'boxes', 'nonMaxSuppression');
-  const $scores = convertToTensor(scores, 'scores', 'nonMaxSuppression');
-
-  const inputs = nonMaxSuppSanityCheck(
-      $boxes, $scores, maxOutputSize, iouThreshold, scoreThreshold,
-      softNmsSigma);
-  maxOutputSize = inputs.maxOutputSize;
-  iouThreshold = inputs.iouThreshold;
-  scoreThreshold = inputs.scoreThreshold;
-  softNmsSigma = inputs.softNmsSigma;
-
-  const attrs = {maxOutputSize, iouThreshold, scoreThreshold, softNmsSigma};
-
-  const result = ENGINE.runKernel(
-                     'NonMaxSuppressionV5', {boxes: $boxes, scores: $scores},
-                     attrs) as Tensor[];
-
-  return {selectedIndices: result[0], selectedScores: result[1]};
-}
-
-/** This is the async version of `nonMaxSuppressionWithScore` */
-async function nonMaxSuppressionWithScoreAsync_(
-    boxes: Tensor2D|TensorLike, scores: Tensor1D|TensorLike,
-    maxOutputSize: number, iouThreshold = 0.5,
-    scoreThreshold = Number.NEGATIVE_INFINITY,
-    softNmsSigma = 0.0): Promise<NamedTensorMap> {
-  const $boxes = convertToTensor(boxes, 'boxes', 'nonMaxSuppressionAsync');
-  const $scores = convertToTensor(scores, 'scores', 'nonMaxSuppressionAsync');
-
-  const inputs = nonMaxSuppSanityCheck(
-      $boxes, $scores, maxOutputSize, iouThreshold, scoreThreshold,
-      softNmsSigma);
-  maxOutputSize = inputs.maxOutputSize;
-  iouThreshold = inputs.iouThreshold;
-  scoreThreshold = inputs.scoreThreshold;
-  softNmsSigma = inputs.softNmsSigma;
-
-  const boxesAndScores = await Promise.all([$boxes.data(), $scores.data()]);
-  const boxesVals = boxesAndScores[0];
-  const scoresVals = boxesAndScores[1];
-
-  const res = nonMaxSuppressionV5(
-      boxesVals, scoresVals, maxOutputSize, iouThreshold, scoreThreshold,
-      softNmsSigma);
-
-  if ($boxes !== boxes) {
-    $boxes.dispose();
-  }
-  if ($scores !== scores) {
-    $scores.dispose();
-  }
-  return res;
-}
 
 /**
  * Extracts crops from the input image tensor and resizes them using bilinear
@@ -231,7 +94,4 @@ function cropAndResize_(
   return res;
 }
 
-export const nonMaxSuppressionAsync = nonMaxSuppressionAsync_;
-export const nonMaxSuppressionWithScore = op({nonMaxSuppressionWithScore_});
-export const nonMaxSuppressionWithScoreAsync = nonMaxSuppressionWithScoreAsync_;
 export const cropAndResize = op({cropAndResize_});
